Refactored library code.
Keeping all of the code in two header files was not
a very maintainable approach. In this first round of
refactoring I have split the code out into separate
header files based on the concept of a single class
per header file as well as separate headers for types.
To maintain compatibility with current users of the
library you are still only required to include
the header autobahn/autobahn.hpp. Other changes I
rolled into this are:

1. Better class naming convention.
2. Fixed indentation to a standard 4 spaces per indent.
3. Drop the use of structs and use classes instead.

The session class is still a bit fat and will need to
be trimmed further moving forward.