Erradicate usage of boost::any
Aside from the performance implications of using boost::any there
was also an issue with adding support for custom types. The addition
of custom types was dependent upon an if-else clause that was doing
string comparisons with RTTI. Needless to say, this obviously wasn't
a performant or scalable approach. So instead of using boost::any as
the variant type we have switched to using msgpack::object. Msgpack
is a serialization protocol/framework however its type msgpack::object
is in fact a variant type much in the same way as boost::any or
boost::variant. However, it was written/designed to be used in a way
that provides good performance and minimizes memory allocaton
overhead.