Make msgpack zones a per result fixture
You have to be careful with msgpack zones. They do not provide a
mechanism to release the objects that they were used to allocate.
As a result, if you do not discard them when the objects that they
were used to allocate go out of scope they can grow indefinitely.
For now we will simply associate a zone with each invocation
result.