Fix use after free issues with call results.
Processing call results happens asynchronously. When we switched
from boost::any to msgpack::object we missed the fact that deep
copies were occring before when setting the call results. However
with msgpack::object, objects are bound to the lifetime of the
zone that they were allocated from. As a result, we need the zone
to live as long as the objects that were allocated from it. This
is acheived by passing the zone into the call result.