Include our own include directory first to avoid using old headers.
It's possible that the Boost include directory is the same as the
one that we previously installed headers in, and we want to make
sure that the local versions from the repo are used instead of
the installed header files. Prioritizing the source directory in
the include order fixes such issues.