Fix examples
I have pruned out a bunch of examples that do not add any
real value. I also got rid of the usage of the boost tcp
resolver as it was causing race conditions. In particular,
the resolve starts thread(s) and calls io_service::run().
This means that the session is being driven by multiple
threads inadvertently which it is not yet designed for.
Plus it produces invalid example code as you can't really
call future::wait() in the context of an example that is
supposed to be showing how to use a session from a single
thread.