Implement wamp components
This commit introduces the concept of components. A component
is simply a wrapper around a session object that provides some
conveniences. This eliminates a significant amount of boilerplate
code required to get a session established. In addition, concrete
component types have also been defined for TCP and Unix Domain
Sockets (UDS).

Fixes: #53