Add abstraction for transports
In order to use this library for larger scale projects with large numbers
of components it needs to support run-time polymorphism for components.
This allows clients/services to be written generically without having to
fixate on specific component types. To achieve this I have removed the
need to have template parameters on the session object. This was done by
abstracting the transport type used by the session. In this manner we can
now have a generic session class that simply takes in the transports
generically as arguments. Now to deploy a precanned service over TCP or
UDS can be done without the need for templatization at the application
level which makes for more reusable/maintainable code. As new transport
types are added/changed they can be utilized without the need for any
extensive code changes.

On the design side, I was given the choice between using either stream
or message based transports. For the first pass I have decided to go with
stream based transports purely for the sake of simplicity and because it
is the smallest incremental step in the right direction. So we get to move
forward in the right direction without having to make any significant code
changes. If message based transports are desired, it is possible to move
in that direction in future commits without having to change any of the
"public" facing api.

Fixes #54